add Old prefix to old filenames, remove New prefix from new filenames
