+++ Sources/SwiftFusion/Inference/JacobianFactor.swift
@@ -1,4 +1,4 @@
-// Copyright 2019 The SwiftFusion Authors. All Rights Reserved.
+// Copyright 2020 The SwiftFusion Authors. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -11,133 +11,97 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-/// A `LinearFactor` that operates like `JacobianFactor` in GTSAM.
-///
-/// Input is a dictionary of `Key` to `Tensor` pairs, and the output is the paired
-/// error vector. Note here that the shapes are not checked.
-///
-/// Interpretation
-/// ================
-/// `Input`: the input values as key-value pairs
-/// `Errors`: the vector `J * x`, where `J` is `m*n` and `x` is `n*1` and returns a m vector
-///
-/// In block form, it looks like the following:
-/// ```
-///                 x1
-/// [ J1 J2 J3 ] *  x2 = [ J1*x1 J2*x2 J3*x3 ]
-///                 x3
-/// ```
-///
-/// Explanation
-/// ================
-/// TODO:
-/// I think both Jacobian and Hessian factors in GTSAM are converted into JacobianFactor in
-/// `GaussianFactorGraph`, so it becomes a question whether we should do the same?
-/// I am considering making `JacobianLikeFactor` a protocol and make `JacobianFactor`
-/// and `HessianFactor` conform to this protocol instead.
-public struct OldJacobianFactor: LinearFactor {
-  
-  // TODO(fan): correct this and add a unit test
-  public func error(_ values: VectorValues) -> ScalarType {
-    ScalarType.zero
+
+import PenguinStructures
+
+/// A Gaussian distribution over the input, represented as a materialized Jacobian matrix and
+/// materialized error vector.
+public struct JacobianFactor<
+  Rows: FixedSizeArray,
+  ErrorVector: EuclideanVectorN
+>: GaussianFactor where Rows.Element: EuclideanVectorN & VariableTuple {
+  public typealias Variables = Rows.Element
+
+  /// The Jacobian matrix, as a fixed size array of rows.
+  ///
+  /// The `jacobian` has one row per element of the `ErrorVector`, and each row is a vector in the
+  /// vector space of adjacent variables. For example, if `ErrorVector == Vector3` and
+  /// `Variables == Tuple2<Vector3, Vector3>`, then `Rows == Array3<Tuple2<Vector3, Vector3>>`. See
+  /// the typealiases below for more examples.
+  public let jacobian: Rows
+
+  /// The error vector.
+  public let error: ErrorVector
+
+  /// The ids of the variables adjacent to this factor.
+  public let edges: Variables.Indices
+
+  /// Creates a Jacobian factor with the given `jacobian`, `error`, and `edges`.
+  init(jacobian: Rows, error: ErrorVector, edges: Variables.Indices) {
+    self.jacobian = jacobian
+    self.error = error
+    self.edges = edges
   }
-  
-  public var dimension: Int {
-    get {
-      jacobians[0].rowCount
+
+  /// Creates a Jacobian factor that linearizes `f` at `x`, and is adjacent to the variables
+  /// identifed by edges.
+  init<Input: Differentiable & DifferentiableVariableTuple>(
+    linearizing f: @differentiable (Input) -> ErrorVector,
+    at x: Input,
+    edges: Input.Indices
+  ) where Input.TangentVector == Variables, Input.TangentIndices == Variables.Indices {
+    let (value, pb) = valueWithPullback(at: x, in: f)
+    let rows = Rows(ErrorVector.standardBasis.lazy.map(pb))
+    self.jacobian = rows
+    self.error = value
+    self.edges = Input.linearized(edges)
+  }
+
+  public func error(at x: Variables) -> Double {
+    return errorVector(at: x).squaredNorm
+  }
+
+  public func errorVector(at x: Variables) -> ErrorVector {
+    return error - errorVector_linearComponent(x)
+  }
+
+  public func errorVector_linearComponent(_ x: Variables) -> ErrorVector {
+    // The compiler isn't able to optimize the closure away if we map `jacobian`, but it is able
+    // to optimize the closure away if we map `jacobian`'s `UnsafeBufferPointer`.
+    jacobian.withUnsafeBufferPointer { rows in
+      ErrorVector(rows.lazy.map { $0.dot(x) })
     }
   }
-  public var keys: Array<Int>
-  public var jacobians: Array<Matrix>
-  public var b: Vector
-  public typealias Output = Error
-  
-  public init (_ key: [Int], _ A: [Matrix], _ b: Vector) {
-    keys = key
-    jacobians = A
-    self.b = b
-  }
-  
-  /// Calculate `J*x`
-  /// Comparable to the `*` operator in GTSAM
-  /// In block form, it looks like the following:
-  /// ```
-  ///                 x1
-  /// [ J1 J2 J3 ] *  x2 = [ J1*x1 J2*x2 J3*x3 ]
-  ///                 x3
-  /// ```
-  static func * (lhs: OldJacobianFactor, rhs: VectorValues) -> Self.Output {
-    var result = Vector(zeros: lhs.dimension)
-    for i in lhs.keys.indices {
-      result += matvec(lhs.jacobians[i], rhs[lhs.keys[i]])
-    }
-    return result
-  }
-  
-  /// Calculate `J^T * e`
-  /// `J^T` is `n*m` and e is `m*1`
-  /// In block form, it looks like the following:
-  /// ```
-  /// J1^T          J1Te
-  /// J2^T  *  e =  J2Te
-  /// J3^T          J3Te
-  /// ```
-  /// However, there exists the possibility that one Value is used multiple times
-  /// At that time we need to add the corresponding blocks
-  public func atr (_ r: Self.Output) -> VectorValues {
-    var result = VectorValues()
-    
-    // No noise model yet
-    // if (model_) model_->whitenInPlace(E);
-    
-    // Just iterate over all A matrices and insert Ai^e into VectorValues
-    for pos in 0..<keys.count {
-      let k = keys[pos]
-      
-      // TODO(fan): add a proper method for searching key
-      if let ind = result._indices[k] {
-        result._values[ind] += matvec(jacobians[pos], transposed: true, r)
-      } else {
-        result.insert(k, matvec(jacobians[pos], transposed: true, r))
+
+  public func errorVector_linearComponent_adjoint(_ y: ErrorVector) -> Variables {
+    // We use `UnsafeBufferPointer`s to avoid forming collections that can't be optimized away.
+    y.withUnsafeBufferPointer { scalars in
+      jacobian.withUnsafeBufferPointer { rows in
+        // We reduce the range `0..<ErrorVector.dimension` instead of `zip(scalars, rows)`, to
+        // avoid forming collections that can't be optimized away.
+        // TODO: This is not getting unrolled after `ErrorVector` is specialized. Convincing the
+        // optimizer to unroll it might speed things up.
+        (0..<ErrorVector.dimension).reduce(into: Variables.zero) { (result, i) in
+          result += scalars[i] * rows[i]
+        }
       }
     }
-    return result
+  }
+
+  public typealias Linearization = Self
+  public func linearized(at x: Variables) -> Self {
+    return self
   }
 }
 
-extension OldJacobianFactor {
-  /// Creates a `OldJacobianFactor` by linearizing the error function `f` at `p`.
-  public init<R: EuclideanVectorN>(
-    of f: @differentiable (Values) -> R,
-    at p: Values
-  ) {
-    // Compute the rows of the jacobian.
-    let (value, pb) = valueWithPullback(at: p, in: f)
-    let rows = R.TangentVector.standardBasis.map { pb($0) }
+/// A Jacobian factor with 1 3-dimensional input and a 3-dimensional error vector.
+public typealias JacobianFactor3x3_1 = JacobianFactor<Array3<Tuple1<Vector3>>, Vector3>
 
-    // Construct empty matrices with the correct shape.
-    assert(rows.count > 0)
-    var matrices = Dictionary<Int, Matrix>(uniqueKeysWithValues: rows[0].keys.map { key in
-      let row = rows[0][key]
-      var matrix = Matrix([], rowCount: 0, columnCount: row.dimension)
-      matrix.reserveCapacity(rows.count * row.dimension)
-      return (key, matrix)
-    })
+/// A Jacobian factor with 2 3-dimensional inputs and a 3-dimensional error vector.
+public typealias JacobianFactor3x3_2 = JacobianFactor<Array3<Tuple2<Vector3, Vector3>>, Vector3>
 
-    // Fill in the matrix entries.
-    for row in rows {
-      for key in row.keys {
-        matrices[key]!.append(row: row[key])
-      }
-    }
+/// A Jacobian factor with 1 6-dimensional input and a 6-dimensional error vector.
+public typealias JacobianFactor6x6_1 = JacobianFactor<Array6<Tuple1<Vector6>>, Vector6>
 
-    // Return the jacobian factor with the matrices and value.
-    let orderedKeys = Array(matrices.keys)
-    self = OldJacobianFactor(
-      orderedKeys,
-      orderedKeys.map { matrices[$0]! },
-      // TODO: remove this negative sign
-      -value.vector
-    )
-  }
-}
+/// A Jacobian factor with 2 6-dimensional inputs and a 6-dimensional error vector.
+public typealias JacobianFactor6x6_2 = JacobianFactor<Array6<Tuple2<Vector6, Vector6>>, Vector6>
